Move useGetOne and useGetList hooks to fetch directory
