+++ packages/ra-core/src/controller/index.ts
@@ -6,8 +6,6 @@ import CreateController from './CreateController';
 import EditController from './EditController';
 import ListController from './ListController';
 import ShowController from './ShowController';
-import useGetOne from './useGetOne';
-import useGetList from './useGetList';
 import useRecordSelection from './useRecordSelection';
 import useVersion from './useVersion';
 export {
@@ -17,8 +15,6 @@ export {
     EditController,
     ListController,
     ShowController,
-    useGetOne,
-    useGetList,
     useRecordSelection,
     useVersion,
 };
