+++ packages/ra-core/src/fetch/index.ts
@@ -7,6 +7,8 @@ import useMutation from './useMutation';
 import useQuery from './useQuery';
 import useQueryWithStore from './useQueryWithStore';
 import withDataProvider from './withDataProvider';
+import useGetOne from './useGetOne';
+import useGetList from './useGetList';
 
 export {
     fetchUtils,
@@ -16,6 +18,8 @@ export {
     useDataProvider,
     useMutation,
     useQuery,
+    useGetOne,
+    useGetList,
     useQueryWithStore,
     withDataProvider,
 };
